app/requirements.txt
fastapi
uvicorn
aiohttp
python-dotenv
slowapi
jinja2
python-multipart
tiktoken
pillow
python-docx
python-pptx
openpyxl
xlrd
PyPDF2
aiofiles
aiosqlite
apscheduler
